Oops, this should be an int.

+++ Library/Tests/Koala/KoalaTests.swift
@@ -106,7 +106,7 @@ final class KoalaTests: XCTestCase {
     XCTAssertEqual(project.stats.backersCount, properties["project_backers_count"] as? Int)
 
     XCTAssertEqual(24 * 15, properties["project_hours_remaining"] as? Int)
-    XCTAssertEqual(60 * 60 * 24 * 30, properties["project_duration"] as? Double)
+    XCTAssertEqual(60 * 60 * 24 * 30, properties["project_duration"] as? Int)
 
     XCTAssertEqual("discovery", properties["ref_tag"] as? String)
     XCTAssertEqual("recommended", properties["referrer_credit"] as? String)
